Fix archival
